IMprove the formatting of listed patches. Rename patches and files to be consistent and sensical
