+++ requirements.txt
@@ -4,4 +4,5 @@ click==7.1.2
 ujson==2.0.3
 requests==2.24.0
 pytest==5.4.3
+countryinfo=0.1.2
 git+https://github.com/paramsingh/pylistenbrainz.git@v0.4.1#egg=pylistenbrainz
